split rpc/traits
