+++ crates/flashblocks/src/lib.rs
@@ -15,5 +15,6 @@ pub mod rpc;
 pub mod state;
 
 pub mod subscription;
+pub mod traits;
 
 pub use pending_blocks::PendingBlocks;
