+++ crates/flashblocks/src/rpc.rs
@@ -13,7 +13,6 @@ use alloy_rpc_types::{
     state::{EvmOverrides, StateOverride, StateOverridesBuilder},
 };
 use alloy_rpc_types_eth::{Filter, Log};
-use arc_swap::Guard;
 use jsonrpsee::{
     core::{RpcResult, async_trait},
     proc_macros::rpc,
@@ -29,53 +28,17 @@ use reth_rpc_eth_api::{
     EthApiTypes, EthFilterApiServer, RpcBlock, RpcReceipt, RpcTransaction,
     helpers::{EthBlocks, EthCall, EthState, EthTransactions, FullEthApi},
 };
-use tokio::{
-    sync::{broadcast, broadcast::error::RecvError},
-    time,
-};
+use tokio::{sync::broadcast::error::RecvError, time};
 use tokio_stream::{StreamExt, wrappers::BroadcastStream};
 use tracing::{debug, trace, warn};
 
-use crate::{PendingBlocks, metrics::Metrics};
+use crate::{
+    metrics::Metrics,
+    traits::{FlashblocksAPI, PendingBlocksAPI},
+};
 
 /// Max configured timeout for `eth_sendRawTransactionSync` in milliseconds.
 pub const MAX_TIMEOUT_SEND_RAW_TX_SYNC_MS: u64 = 6_000;
-
-/// Core API for accessing flashblock state and data.
-pub trait FlashblocksAPI {
-    /// Retrieves the pending blocks.
-    fn get_pending_blocks(&self) -> Guard<Option<Arc<PendingBlocks>>>;
-
-    /// Subscribes to flashblock updates.
-    fn subscribe_to_flashblocks(&self) -> broadcast::Receiver<Arc<PendingBlocks>>;
-}
-
-/// API for accessing pending blocks data.
-pub trait PendingBlocksAPI {
-    /// Get the canonical block number on top of which all pending state is built
-    fn get_canonical_block_number(&self) -> BlockNumberOrTag;
-
-    /// Get the pending transactions count for an address
-    fn get_transaction_count(&self, address: Address) -> U256;
-
-    /// Retrieves the current block. If `full` is true, includes full transaction details.
-    fn get_block(&self, full: bool) -> Option<RpcBlock<Optimism>>;
-
-    /// Gets transaction receipt by hash.
-    fn get_transaction_receipt(&self, tx_hash: TxHash) -> Option<RpcReceipt<Optimism>>;
-
-    /// Gets transaction details by hash.
-    fn get_transaction_by_hash(&self, tx_hash: TxHash) -> Option<RpcTransaction<Optimism>>;
-
-    /// Gets balance for an address. Returns None if address not updated in flashblocks.
-    fn get_balance(&self, address: Address) -> Option<U256>;
-
-    /// Gets the state overrides for the pending blocks
-    fn get_state_overrides(&self) -> Option<StateOverride>;
-
-    /// Gets logs from pending state matching the provided filter.
-    fn get_pending_logs(&self, filter: &Filter) -> Vec<Log>;
-}
 
 /// Eth API override trait for flashblocks integration.
 #[cfg_attr(not(test), rpc(server, namespace = "eth"))]
